app/requirements.txt
mysql-connector-python==8.0.28
redis==4.0.0